backend/requirements.txt

numpy==1.16.4
pandas==0.24.2
python-dotenv==0.10.3
pymysql==0.9.3
Flask==1.1.1
flask-cors==3.0.7
SQLAlchemy==1.3.5
